feat: implement vuex store and saving in local storage
