+++ src/components/organisms/task-list/TaskList.vue
@@ -1,7 +1,7 @@
 <template>
-  <div class="w-full flex flex-col gap-4">
+  <div v-if="$store.state.tasks.length" class="w-full flex flex-col gap-4">
     <task-item
-      v-for="(task, index) of tasks"
+      v-for="(task, index) in $store.state.tasks"
       :id="index"
       :key="index"
       :value="task.text"
@@ -11,25 +11,13 @@
 </template>
 
 <script lang="ts">
-import { defineComponent, type PropType } from 'vue';
+import { defineComponent } from 'vue';
 import TaskItem from '../../molecules/task-item/TaskItem.vue';
-
-export type Task = {
-  text: string;
-  completed: boolean;
-};
 
 export default defineComponent({
   name: 'TaskList',
   components: {
     TaskItem,
-  },
-
-  props: {
-    tasks: {
-      type: Array as PropType<Task[]>,
-      default: (): Task[] => [],
-    },
   },
 });
 </script>
