+++ src/main.ts
@@ -1,6 +1,6 @@
 import Vue from 'vue';
 import App from './App.vue';
-import store from './store';
+import store from './store/store';
 
 Vue.config.productionTip = false;
